Improve the error message when the wrong number of arguments are provided
Currently when a user provides the wrong number of arguments, they receive a
message saying `echo is not implemented.`.

This message is confusing because it makes the user feel like they typed in the
wrong function name, when in reality they just provided the wrong number of
arguments.

This is a first draft at a fix - feedback welcome!